修改消费者注册模式、修改部分类命名、添加线程配置等功能

+++ library/src/main/java/cn/knowbox/book/alimq/producer/RocketMqOrderType.java
@@ -29,7 +29,7 @@ public enum RocketMqOrderType {
         this.type = code;
     }
 
-    public static RocketMqOrderType valueOf(byte type) {
+    public static RocketMqOrderType valueOf(Integer type) {
         for (RocketMqOrderType item : RocketMqOrderType.values()) {
             if (item.getType() == type) {
                 return item;
